Voting: use aragonAPI for React + React Hooks refactoring (#717)
The app is now using aragonAPI for React, and has also been a bit rearranged
to take advantage of using React Hooks. The idea is to keep using the Voting app
as an experiment to find and validate patterns related to React Hooks that we could use.

A new file, `app-logic.js`, contains most of the logic to be used by `App.js`, by exporting
`AppLogicProvider` and `useAppLogic()`. This is meant to be the pattern used by other apps
in the future (providing that we validate this idea).

Other changes:

- VotingCard now skips any unnecessary update.
- The vote texts (description and metadata) are now a separate component rather than
  being pre-rendered (that way it is easier to skip updates).